Initial commit for C++ Kalman Filter implementation.
The filter is implemented as a class using the Eigen linear algebra
library. At the moment two versions are implemented, one using
dynamic sized matrices, the other fixed size.
Example usage can be found in the test-script.cpp file. To run it
you need to have the Eigen header file in your path.